BraveWitch - early game creation 0.06 (fight options)

+++ src/main/java/AbleToAttack.java
@@ -1,5 +1,5 @@
 public interface AbleToAttack {
 
-Integer attack();
+double attack();
 
 }
